requirements.txt
certifi==2023.7.22
charset-normalizer==3.3.1
contourpy==1.1.1
customtkinter==5.2.1
cycler==0.12.1
darkdetect==0.8.0
fonttools==4.43.1
idna==3.4
kiwisolver==1.4.5
matplotlib==3.8.0
numpy==1.26.1
packaging==23.2
Pillow==10.1.0
pyparsing==3.1.1
pyserial==3.5
python-dateutil==2.8.2
requests==2.31.0
six==1.16.0
urllib3==2.0.7
